Move macros and helpers to fn_utils
- Remove obsolete ctx parameter
- make ARGSEL(S) macro specific
