+++ src/context.cpp
@@ -29,6 +29,7 @@
 #include "sass2scss.h"
 #include "prelexer.hpp"
 #include "emitter.hpp"
+#include "fn_utils.hpp"
 
 namespace Sass {
   using namespace Constants;
